[REF] Changed int id to long id

+++ src/com/excilys/computerdatabase/main/java/dao/ComputerDAO.java
@@ -60,14 +60,14 @@ public enum ComputerDAO {
 		return new Page<Computer>(COMPUTERS_PER_PAGE, offset, cpuList);
 	}
 
-	public Computer getComputer(int id) throws Exception {
+	public Computer getComputer(long id) throws Exception {
 
 		Connection connection = null;
 		Computer cpu = null;
 		try {
 			connection = getConnection();
 			PreparedStatement st = connection.prepareStatement(FIND_BY_ID);
-			st.setInt(1, id);
+			st.setLong(1, id);
 			ResultSet rs = st.executeQuery();
 
 			while (rs.next()) {
@@ -116,7 +116,7 @@ public enum ComputerDAO {
 			} else {
 				st.setInt(4, cpu.getCompany().getId());
 			}
-			st.setInt(5, cpu.getId());
+			st.setLong(5, cpu.getId());
 			return st.executeUpdate();
 		} catch (SQLException e) {
 			throw new Exception(e.getMessage(), e);
@@ -125,12 +125,12 @@ public enum ComputerDAO {
 		}
 	}
 
-	public int deleteComputer(int id) throws Exception {
+	public int deleteComputer(long id) throws Exception {
 		Connection connection = null;
 		try {
 			connection = getConnection();
 			PreparedStatement st = connection.prepareStatement(DELETE);
-			st.setInt(1, id);
+			st.setLong(1, id);
 			return st.executeUpdate();
 		} catch (SQLException e) {
 			throw new Exception(e.getMessage(), e);
